Выделяет компонент 'Catalog'.
