+++ src/components/main-page/main-page.jsx
@@ -1,35 +1,7 @@
 import PropTypes from "prop-types";
 import React from "react";
-import {CatalogGenresItem} from "../catalog-genres-item/catalog-genres-item.jsx";
-import {CatalogMore} from "../catalog-more/catalog-more.jsx";
-import {SmallMovieCard} from "../small-movie-card/small-movie-card.jsx";
+import {Catalog} from "../catalog/catalog.jsx";
 import {MovieCard} from "../movie-card/movie-card.jsx";
-
-
-const genreNames = [
-  `All genres`,
-  `Comedies`,
-  `Crime`,
-  `Documentary`,
-  `Dramas`,
-  `Horror`,
-  `Kids & Family`,
-  `Romance`,
-  `Sci-Fi`,
-  `Thrillers`,
-];
-
-const handleGenreClick = (event) => {
-  event.preventDefault();
-};
-
-const handleSmallMovieCardClick = (event) => {
-  event.preventDefault();
-};
-
-const handleCatalogButtonClick = (event) => {
-  event.preventDefault();
-};
 
 
 export const MainPage = (props) => {
@@ -69,35 +41,7 @@ export const MainPage = (props) => {
       <MovieCard {...promoMovie} />
 
       <div className="page-content">
-        <section className="catalog">
-          <h2 className="catalog__title visually-hidden">Catalog</h2>
-
-          <ul className="catalog__genres-list">
-            {
-              genreNames.map((genreName, index) =>
-                <CatalogGenresItem
-                  key={genreName + index}
-                  genreName={genreName}
-                  onGenreClick={handleGenreClick}
-                />
-              )
-            }
-          </ul>
-
-          <div className="catalog__movies-list">
-            {
-              movieTitles.map((movieTitle, index) =>
-                <SmallMovieCard
-                  key={movieTitle + index}
-                  movieTitle={movieTitle}
-                  onSmallMovieCardClick={handleSmallMovieCardClick}
-                />
-              )
-            }
-          </div>
-
-          <CatalogMore onClick={handleCatalogButtonClick} />
-        </section>
+        <Catalog movieTitles={movieTitles} />
 
         <footer className="page-footer">
           <div className="logo">
